refactor: 직무 패키지 변경
