+++ domain/src/main/java/com/pinback/domain/user/entity/User.java
@@ -5,7 +5,7 @@ import java.time.LocalTime;
 import java.util.UUID;
 
 import com.pinback.domain.common.BaseEntity;
-import com.pinback.domain.user.enums.Job;
+import com.pinback.domain.common.enums.Job;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
