+++ infrastructure/src/main/java/com/pinback/infrastructure/article/service/ArticleGetService.java
@@ -20,8 +20,8 @@ import com.pinback.application.common.exception.ArticleNotFoundException;
 import com.pinback.application.common.exception.ArticleNotOwnedException;
 import com.pinback.domain.article.entity.Article;
 import com.pinback.domain.category.entity.Category;
+import com.pinback.domain.common.enums.Job;
 import com.pinback.domain.user.entity.User;
-import com.pinback.domain.user.enums.Job;
 import com.pinback.infrastructure.article.repository.ArticleRepository;
 import com.pinback.infrastructure.article.repository.dto.ArticleCountInfoV3;
 import com.pinback.infrastructure.article.repository.dto.ArticleWithCountV3;
