+++ infrastructure/src/main/java/com/pinback/infrastructure/user/service/UserUpdateService.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.pinback.application.user.port.out.UserUpdateServicePort;
+import com.pinback.domain.common.enums.Job;
 import com.pinback.domain.user.entity.User;
-import com.pinback.domain.user.enums.Job;
 import com.pinback.infrastructure.user.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
